Remove push notification subscription management responsibility from SettingService
The SettingService used to be created very soon after launch, so it would catch the new registration token. It is now only created when needed by the Settings view controller, which makes it a bad place to keep the listener responsible for updating the current subscriptions with the new device token.

Also, update all subscriptions on launch, rather than each account’s subscriptions only when they become the active account.

Contributes to IOS-384